LayoutTests/css3/flexbox/css-properties-expected.txt

Tests being able to set the display to -webkit-flex and -webkit-inline-flex.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS flexbox.style.display is "-webkit-flex"
PASS flexbox.style.display is "inline"
PASS flexbox.style.display is "-webkit-inline-flex"
PASS flexbox.style.display is "-webkit-inline-flex"
PASS flexbox.style.display is "block"
PASS flexitem.style.webkitFlexOrder is ""
PASS flexitem.style.webkitFlexOrder is "2"
PASS flexitem.style.webkitFlexOrder is "-1"
PASS flexitem.style.webkitFlexOrder is "0"
PASS flexitem.style.webkitFlexOrder is "0"
PASS flexitem.style.webkitFlexOrder is "0"
PASS flexitem.style.webkitFlexOrder is ""
PASS flexbox.style.webkitFlexPack is ""
PASS window.getComputedStyle(flexbox, null).webkitFlexPack is "start"
PASS flexbox.style.webkitFlexPack is ""
PASS flexbox.style.webkitFlexPack is "start"
PASS window.getComputedStyle(flexbox, null).webkitFlexPack is "start"
PASS flexbox.style.webkitFlexPack is "end"
PASS window.getComputedStyle(flexbox, null).webkitFlexPack is "end"
PASS flexbox.style.webkitFlexPack is "center"
PASS window.getComputedStyle(flexbox, null).webkitFlexPack is "center"
PASS flexbox.style.webkitFlexPack is "justify"
PASS window.getComputedStyle(flexbox, null).webkitFlexPack is "justify"
PASS flexbox.style.webkitFlexPack is ""
PASS window.getComputedStyle(flexbox, null).webkitFlexPack is "start"
PASS flexbox.style.webkitFlexItemAlign is ""
PASS window.getComputedStyle(flexbox, null).webkitFlexItemAlign is "stretch"
PASS window.getComputedStyle(document.documentElement, null).webkitFlexItemAlign is "stretch"
PASS flexbox.style.webkitFlexItemAlign is ""
PASS flexbox.style.webkitFlexItemAlign is "auto"
PASS window.getComputedStyle(flexbox, null).webkitFlexItemAlign is "stretch"
PASS flexbox.style.webkitFlexItemAlign is "start"
PASS window.getComputedStyle(flexbox, null).webkitFlexItemAlign is "start"
PASS flexbox.style.webkitFlexItemAlign is "end"
PASS window.getComputedStyle(flexbox, null).webkitFlexItemAlign is "end"
PASS flexbox.style.webkitFlexItemAlign is "center"
PASS window.getComputedStyle(flexbox, null).webkitFlexItemAlign is "center"
PASS flexbox.style.webkitFlexItemAlign is "stretch"
PASS window.getComputedStyle(flexbox, null).webkitFlexItemAlign is "stretch"
PASS flexbox.style.webkitFlexItemAlign is "baseline"
PASS window.getComputedStyle(flexbox, null).webkitFlexItemAlign is "baseline"
PASS flexbox.style.webkitFlexItemAlign is ""
PASS window.getComputedStyle(flexbox, null).webkitFlexItemAlign is "stretch"
PASS flexbox.style.webkitFlexAlign is ""
PASS window.getComputedStyle(flexbox, null).webkitFlexAlign is "stretch"
PASS window.getComputedStyle(flexitem, null).webkitFlexItemAlign is "stretch"
PASS flexbox.style.webkitFlexAlign is ""
PASS flexbox.style.webkitFlexAlign is ""
PASS window.getComputedStyle(flexbox, null).webkitFlexAlign is "stretch"
PASS window.getComputedStyle(flexitem, null).webkitFlexItemAlign is "stretch"
PASS flexbox.style.webkitFlexAlign is "start"
PASS window.getComputedStyle(flexbox, null).webkitFlexAlign is "start"
PASS window.getComputedStyle(flexitem, null).webkitFlexItemAlign is "start"
PASS flexbox.style.webkitFlexAlign is "end"
PASS window.getComputedStyle(flexbox, null).webkitFlexAlign is "end"
PASS window.getComputedStyle(flexitem, null).webkitFlexItemAlign is "end"
PASS flexbox.style.webkitFlexAlign is "center"
PASS window.getComputedStyle(flexbox, null).webkitFlexAlign is "center"
PASS window.getComputedStyle(flexitem, null).webkitFlexItemAlign is "center"
PASS flexbox.style.webkitFlexAlign is "stretch"
PASS window.getComputedStyle(flexbox, null).webkitFlexAlign is "stretch"
PASS window.getComputedStyle(flexitem, null).webkitFlexItemAlign is "stretch"
PASS flexbox.style.webkitFlexAlign is "baseline"
PASS window.getComputedStyle(flexbox, null).webkitFlexAlign is "baseline"
PASS window.getComputedStyle(flexitem, null).webkitFlexItemAlign is "baseline"
PASS flexbox.style.webkitFlexAlign is ""
PASS window.getComputedStyle(flexbox, null).webkitFlexAlign is "stretch"
PASS window.getComputedStyle(flexitem, null).webkitFlexItemAlign is "stretch"
PASS flexbox.style.webkitFlexAlign is ""
PASS window.getComputedStyle(flexbox, null).webkitFlexAlign is "stretch"
PASS window.getComputedStyle(flexitem, null).webkitFlexItemAlign is "stretch"
PASS window.getComputedStyle(detachedFlexbox, null).webkitFlexItemAlign is ""
PASS window.getComputedStyle(detachedFlexItem, null).webkitFlexItemAlign is ""
PASS flexbox.style.webkitFlexDirection is ""
PASS window.getComputedStyle(flexbox, null).webkitFlexDirection is "row"
PASS flexbox.style.webkitFlexDirection is ""
PASS flexbox.style.webkitFlexDirection is "row"
PASS window.getComputedStyle(flexbox, null).webkitFlexDirection is "row"
PASS flexbox.style.webkitFlexDirection is "row-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexDirection is "row-reverse"
PASS flexbox.style.webkitFlexDirection is "column"
PASS window.getComputedStyle(flexbox, null).webkitFlexDirection is "column"
PASS flexbox.style.webkitFlexDirection is "column-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexDirection is "column-reverse"
PASS flexbox.style.webkitFlexWrap is ""
PASS window.getComputedStyle(flexbox, null).webkitFlexWrap is "none"
PASS flexbox.style.webkitFlexWrap is ""
PASS flexbox.style.webkitFlexWrap is "none"
PASS window.getComputedStyle(flexbox, null).webkitFlexWrap is "none"
PASS flexbox.style.webkitFlexWrap is "wrap"
PASS window.getComputedStyle(flexbox, null).webkitFlexWrap is "wrap"
PASS flexbox.style.webkitFlexWrap is "wrap-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexWrap is "wrap-reverse"
PASS flexbox.style.webkitFlexFlow is ""
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row none"
PASS flexbox.style.webkitFlexFlow is ""
PASS flexbox.style.webkitFlexFlow is ""
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row none"
PASS flexbox.style.webkitFlexFlow is ""
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row none"
PASS flexbox.style.webkitFlexFlow is "none"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row none"
PASS flexbox.style.webkitFlexFlow is "none"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row none"
PASS flexbox.style.webkitFlexFlow is "wrap"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row wrap"
PASS flexbox.style.webkitFlexFlow is "wrap"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row wrap"
PASS flexbox.style.webkitFlexFlow is "wrap-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row wrap-reverse"
PASS flexbox.style.webkitFlexFlow is "wrap-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row wrap-reverse"
PASS flexbox.style.webkitFlexFlow is "row"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row none"
PASS flexbox.style.webkitFlexFlow is "row"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row none"
PASS flexbox.style.webkitFlexFlow is "row none"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row none"
PASS flexbox.style.webkitFlexFlow is "row none"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row none"
PASS flexbox.style.webkitFlexFlow is "row wrap"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row wrap"
PASS flexbox.style.webkitFlexFlow is "row wrap"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row wrap"
PASS flexbox.style.webkitFlexFlow is "row wrap-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row wrap-reverse"
PASS flexbox.style.webkitFlexFlow is "row wrap-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row wrap-reverse"
PASS flexbox.style.webkitFlexFlow is "row-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row-reverse none"
PASS flexbox.style.webkitFlexFlow is "row-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row-reverse none"
PASS flexbox.style.webkitFlexFlow is "row-reverse none"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row-reverse none"
PASS flexbox.style.webkitFlexFlow is "row-reverse none"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row-reverse none"
PASS flexbox.style.webkitFlexFlow is "row-reverse wrap"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row-reverse wrap"
PASS flexbox.style.webkitFlexFlow is "row-reverse wrap"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row-reverse wrap"
PASS flexbox.style.webkitFlexFlow is "row-reverse wrap-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row-reverse wrap-reverse"
PASS flexbox.style.webkitFlexFlow is "row-reverse wrap-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row-reverse wrap-reverse"
PASS flexbox.style.webkitFlexFlow is "column"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column none"
PASS flexbox.style.webkitFlexFlow is "column"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column none"
PASS flexbox.style.webkitFlexFlow is "column none"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column none"
PASS flexbox.style.webkitFlexFlow is "column none"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column none"
PASS flexbox.style.webkitFlexFlow is "column wrap"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column wrap"
PASS flexbox.style.webkitFlexFlow is "column wrap"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column wrap"
PASS flexbox.style.webkitFlexFlow is "column wrap-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column wrap-reverse"
PASS flexbox.style.webkitFlexFlow is "column wrap-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column wrap-reverse"
PASS flexbox.style.webkitFlexFlow is "column-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column-reverse none"
PASS flexbox.style.webkitFlexFlow is "column-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column-reverse none"
PASS flexbox.style.webkitFlexFlow is "column-reverse none"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column-reverse none"
PASS flexbox.style.webkitFlexFlow is "column-reverse none"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column-reverse none"
PASS flexbox.style.webkitFlexFlow is "column-reverse wrap"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column-reverse wrap"
PASS flexbox.style.webkitFlexFlow is "column-reverse wrap"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column-reverse wrap"
PASS flexbox.style.webkitFlexFlow is "column-reverse wrap-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column-reverse wrap-reverse"
PASS flexbox.style.webkitFlexFlow is "column-reverse wrap-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column-reverse wrap-reverse"
PASS flexbox.style.webkitFlexFlow is ""
PASS flexbox.style.webkitFlexFlow is ""
PASS flexbox.style.webkitFlexFlow is ""
PASS flexbox.style.webkitFlexFlow is ""
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row none"
PASS flexbox.style.webkitFlexFlow is "column"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column none"
PASS flexbox.style.webkitFlexFlow is "column wrap"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column wrap"
PASS flexbox.style.webkitFlexFlow is "row-reverse wrap-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row-reverse wrap-reverse"
PASS flexbox.style.webkitFlexLinePack is ""
PASS window.getComputedStyle(flexbox, null).webkitFlexLinePack is "stretch"
PASS flexbox.style.webkitFlexLinePack is "start"
PASS window.getComputedStyle(flexbox, null).webkitFlexLinePack is "start"
PASS flexbox.style.webkitFlexLinePack is "end"
PASS window.getComputedStyle(flexbox, null).webkitFlexLinePack is "end"
PASS flexbox.style.webkitFlexLinePack is "center"
PASS window.getComputedStyle(flexbox, null).webkitFlexLinePack is "center"
PASS flexbox.style.webkitFlexLinePack is "justify"
PASS window.getComputedStyle(flexbox, null).webkitFlexLinePack is "justify"
PASS flexbox.style.webkitFlexLinePack is "distribute"
PASS window.getComputedStyle(flexbox, null).webkitFlexLinePack is "distribute"
PASS flexbox.style.webkitFlexLinePack is "stretch"
PASS window.getComputedStyle(flexbox, null).webkitFlexLinePack is "stretch"
PASS flexbox.style.webkitFlexLinePack is ""
PASS window.getComputedStyle(flexbox, null).webkitFlexLinePack is "stretch"
PASS flexbox.style.webkitFlexLinePack is ""
PASS window.getComputedStyle(flexbox, null).webkitFlexLinePack is "stretch"
PASS successfullyParsed is true

TEST COMPLETE
